asyncua client and server working (ua.VariantType.Double fixed)

+++ digital_twins/opc-ua-electroingenium/client-asyncua.py
@@ -67,7 +67,7 @@ Read node value from OPCUA
         # client_node.set_value(value)
         
         ## Physic PLC
-        data_value = ua.DataValue(ua.Variant(value, ua.VariantType.Float))
+        data_value = ua.DataValue(ua.Variant(value, ua.VariantType.Double))
         await client_node.set_data_value(data_value)
 
     async def opcua_write_values(self, node_values:dict):
